nodejs and compiler based JSX tranforms and ESM support to produce SSR HTML
