+++ package.json
@@ -27,8 +27,8 @@
     "test:tdd": "npm run test --watch",
     "dist": "rollup -c rollup.config.js",
     "prepublishOnly": "npm run clean && npm run dist",
-    "sandbox": "node ./sandbox.js",
-    "sandbox:develop": "concurrently \"nodemon --watch src --watch test -e js,md,css,html,jsx ./sandbox.js\" \"http-server ./sandbox --open\"",
+    "sandbox": "node --trace-warnings --experimental-loader ./src/jsx-loader.js ./sandbox.js",
+    "sandbox:develop": "concurrently \"nodemon --watch src --watch test -e js,md,css,html,jsx --trace-warnings --experimental-loader ./src/jsx-loader.js ./sandbox.js\" \"http-server ./sandbox --open\"",
     "sandbox:serve": "npm run sandbox && http-server ./sandbox --open"
   },
   "dependencies": {
